client/controller: Move makeLayoutMiddleware to shared.js
