+++ client/controller/shared.js
@@ -1,5 +1,32 @@
-import { setSection as setSectionAction } from 'state/ui/actions';
+/**
+ * External dependencies
+ */
+import React from 'react';
 import noop from 'lodash/noop';
+
+/**
+ * Internal dependencies
+ */
+import { setSection as setSectionAction } from 'state/ui/actions';
+import { getCurrentUser } from 'state/current-user/selectors';
+
+export function makeLayoutMiddleware( LayoutComponent ) {
+	return ( context, next ) => {
+		const { store, primary, secondary, tertiary } = context;
+
+		// On server, only render LoggedOutLayout when logged-out.
+		if ( ! context.isServerSide || getCurrentUser( context.store.getState() ) ) {
+			context.layout = (
+				<LayoutComponent store={ store }
+					primary={ primary }
+					secondary={ secondary }
+					tertiary={ tertiary }
+				/>
+			);
+		}
+		next();
+	};
+}
 
 export function setSection( section ) {
 	return ( context, next = noop ) => {
